# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Product:
    properties:
      code_value:
        type: string
      expiration:
        type: string
      id:
        type: string
      is_published:
        type: boolean
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a api for Products.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Products API
  version: "1.0"
paths:
  /ping:
    get:
      description: Ping for testing de conection
      produces:
      - application/json
      responses:
        "200":
          description: OK
  /product:
    get:
      description: Return list of all the products in the db
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "500":
          description: Internal Server Error
      summary: Get all the products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product in the db
      parameters:
      - description: TOKEN_ENV
        in: header
        name: token
        required: true
        type: string
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Post new product
      tags:
      - Products
  /product/{id}:
    delete:
      description: Delete a product in the db
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "404":
          description: Product Not Found
      summary: Delete product
      tags:
      - Products
    get:
      description: Return a product in the db
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "404":
          description: Product Not Found
      summary: Get product by id
      tags:
      - Products
    put:
      description: Update a product in the db
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
        "404":
          description: Product Not Found
      summary: Update product
      tags:
      - Products
swagger: "2.0"